refactor: moved source code files to src
New organization of the source code files
