+++ src/guiElements/src/inputs/colorPicker.py
